requirements.txt

    fastapi 
    uvicorn 
    pillow 
    python-dotenv 
    google-generativeai
    langchain-google-genai 
    langchain-community
    langchain 
    tiktoken
    chromadb
    beautifulsoup4
    python-multipart